Move SHMContext into zenoh_utils
