+++ rmw_zenoh_cpp/src/detail/rmw_subscription_data.hpp
@@ -33,7 +33,6 @@
 #include "liveliness_utils.hpp"
 #include "message_type_support.hpp"
 #include "attachment_helpers.hpp"
-#include "shm_context.hpp"
 #include "type_support_common.hpp"
 #include "zenoh_utils.hpp"
